First pass of JPF code: CustomCameraManager.update
Mostly just porting JPF code to UMM loadstyle. Some syntatic differences
but no major changes as of yet. The biggest change is that I'm not
reaching for CameraManager fields if I can avoid it. In most cases, they
can be stored on the CustomCameraManager class instead.

This currently committed code only hands cinematics and the case where
the camera is unlocked. Locked camera but not cinematic to come.